進化Nodeのコンポーネント化
