+++ src/app/pokemon/[id]/page.tsx
@@ -10,14 +10,6 @@ import {
   Weight,
   ChevronLeft,
   ChevronRight,
-  ArrowUp,
-  ArrowRight,
-  ArrowDown,
-  ArrowLeft,
-  ArrowUpRight,
-  ArrowDownRight,
-  ArrowDownLeft,
-  ArrowUpLeft,
 } from "lucide-react";
 import {
   PokemonData,
@@ -25,8 +17,6 @@ import {
   PokemonEvolutionNode,
 } from "@/src/domain/pokemon/pokemon";
 import { pokemonTypeStyleDictionary } from "@/src/domain/pokemon/pokemonTypeStyle";
-// import RadarChartComponent from "@/src/components/ui/RaderChartComponent";
-// import { fetchPokemonStatus } from "@/src/components/ui/fetchPokemonStatus";
 import { PageHeader } from "@/src/components/ui/PageHeader";
 import { BaseStatBarChart } from "@/src/components/ui/BaseStatBarChart";
 import { TypeEffectivenessSection } from "@/src/components/ui/TypeEffectivenessSection";
@@ -34,6 +24,7 @@ import { TypeBadge } from "@/src/components/ui/TypeBadge";
 import { RolingBollAnimation } from "@/src/components/ui/RolingBollAnimation";
 import { Button } from "@/src/components/ui/button";
 import { RarityBadge } from "@/src/components/ui/RarityBadge";
+import {EvolutionTreeNode} from "@/src/components/ui/EvolutionTreeNode"
 
 const ExampleDrawer = dynamic(
   () => import("@/src/components/ui/DisplayExplain"),
@@ -56,10 +47,6 @@ export default function BlogPostPage({ params }: { params: Promise<Params> }) {
   useEffect(() => {
     const fetchPokemon = async () => {
       const response = await fetch(`/api/pokemon/${id}`);
-      // if (!response.ok) {
-      //   setNotFound(true);
-      //   return;
-      // }
       const data = await response.json();
       setData(data.pokemon);
       setMaxPokeApiId(data.maxPokeApiId);
@@ -329,188 +316,3 @@ export default function BlogPostPage({ params }: { params: Promise<Params> }) {
   );
 }
 
-// 分岐先を中心から上下左右・斜めのコンパス状に配置する順序
-// (イーブイの8進化のように多分岐の場合に使う)。矢印は中心に近い側に置き、
-// 分岐先の方向を指すようにする
-const HUB_POSITIONS = [
-  {
-    grid: "col-start-2 row-start-1",
-    Icon: ArrowUp,
-    dir: "col",
-    arrowFirst: false,
-  }, // 上
-  {
-    grid: "col-start-3 row-start-2",
-    Icon: ArrowRight,
-    dir: "row",
-    arrowFirst: true,
-  }, // 右
-  {
-    grid: "col-start-2 row-start-3",
-    Icon: ArrowDown,
-    dir: "col",
-    arrowFirst: true,
-  }, // 下
-  {
-    grid: "col-start-1 row-start-2",
-    Icon: ArrowLeft,
-    dir: "row",
-    arrowFirst: false,
-  }, // 左
-  {
-    grid: "col-start-3 row-start-1",
-    Icon: ArrowUpRight,
-    dir: "col",
-    arrowFirst: false,
-  }, // 右上
-  {
-    grid: "col-start-3 row-start-3",
-    Icon: ArrowDownRight,
-    dir: "col",
-    arrowFirst: true,
-  }, // 右下
-  {
-    grid: "col-start-1 row-start-3",
-    Icon: ArrowDownLeft,
-    dir: "col",
-    arrowFirst: true,
-  }, // 左下
-  {
-    grid: "col-start-1 row-start-1",
-    Icon: ArrowUpLeft,
-    dir: "col",
-    arrowFirst: false,
-  }, // 左上
-] as const;
-
-function EvolutionNodeButton({
-  node,
-  isCurrent,
-}: {
-  node: PokemonEvolutionNode;
-  isCurrent: boolean;
-}) {
-  const router = useRouter();
-  return (
-    <button
-      type="button"
-      onClick={() => router.push(`/pokemon/${node.pokeApiId}`)}
-      className={`flex cursor-pointer flex-col items-center gap-1 rounded-2xl border px-3 py-2 transition-colors duration-150 ${
-        isCurrent
-          ? "border-primary bg-primary/10 shadow-sm"
-          : "border-border bg-card hover:bg-accent"
-      }`}
-    >
-      <div className="flex h-14 w-14 items-center justify-center rounded-full bg-muted">
-        {node.imageUrl && (
-          <Image
-            src={node.imageUrl}
-            alt={node.japaneseName}
-            width={48}
-            height={48}
-          />
-        )}
-      </div>
-      <div
-        className={`text-xs font-bold ${
-          isCurrent ? "text-primary" : "text-card-foreground"
-        }`}
-      >
-        {node.japaneseName}
-      </div>
-    </button>
-  );
-}
-
-function EvolutionTreeNode({
-  node,
-  currentPokedexId,
-}: {
-  node: PokemonEvolutionNode;
-  currentPokedexId: number;
-}) {
-  const isCurrent = node.pokedexId === currentPokedexId;
-
-  if (node.evolvesTo.length === 0) {
-    return <EvolutionNodeButton node={node} isCurrent={isCurrent} />;
-  }
-
-  // イーブイ(8分岐)のように多分岐の場合は、中心から上下左右に広げるレイアウトにする
-  if (node.evolvesTo.length >= 5) {
-    const hubChildren = node.evolvesTo.slice(0, HUB_POSITIONS.length);
-    const overflowChildren = node.evolvesTo.slice(HUB_POSITIONS.length);
-    return (
-      <div className="flex flex-col items-center gap-4">
-        <div className="grid grid-cols-3 grid-rows-3 place-items-center gap-3">
-          <div className="col-start-2 row-start-2">
-            <EvolutionNodeButton node={node} isCurrent={isCurrent} />
-          </div>
-          {hubChildren.map((child, index) => {
-            const pos = HUB_POSITIONS[index];
-            const Icon = pos.Icon;
-            const arrow = (
-              <Icon size={16} className="shrink-0 text-muted-foreground" />
-            );
-            const nodeEl = (
-              <EvolutionTreeNode
-                node={child}
-                currentPokedexId={currentPokedexId}
-              />
-            );
-            return (
-              <div
-                key={child.pokedexId}
-                className={`flex items-center justify-center gap-1 ${
-                  pos.dir === "row" ? "flex-row" : "flex-col"
-                } ${pos.grid}`}
-              >
-                {pos.arrowFirst ? (
-                  <>
-                    {arrow}
-                    {nodeEl}
-                  </>
-                ) : (
-                  <>
-                    {nodeEl}
-                    {arrow}
-                  </>
-                )}
-              </div>
-            );
-          })}
-        </div>
-        {overflowChildren.length > 0 && (
-          <div className="flex flex-wrap justify-center gap-2">
-            {overflowChildren.map((child) => (
-              <EvolutionTreeNode
-                key={child.pokedexId}
-                node={child}
-                currentPokedexId={currentPokedexId}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex items-center gap-2">
-      <EvolutionNodeButton node={node} isCurrent={isCurrent} />
-      <div className="flex flex-col gap-2">
-        {node.evolvesTo.map((child) => (
-          <div key={child.pokedexId} className="flex items-center gap-2">
-            <ChevronRight
-              size={18}
-              className="shrink-0 text-muted-foreground"
-            />
-            <EvolutionTreeNode
-              node={child}
-              currentPokedexId={currentPokedexId}
-            />
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
